add unidentifiedPackages attribute to the tree. extract dependency-tree code to a new component
